refactor: separate game engine and utilities
