+++ src/core/constants.js
@@ -1,0 +1,16 @@
+export const RANKS = [
+  "2",
+  "3",
+  "4",
+  "5",
+  "6",
+  "7",
+  "8",
+  "9",
+  "10",
+  "J",
+  "Q",
+  "K",
+  "A",
+];
+export const SUITS = ["C", "D", "H", "S"]; // Clubs, Diamonds, Hearts, Spades
